rename x/feedist to x/distribution
