+++ x/distribution/msg_test.go
@@ -1,0 +1,1 @@
+package distribution
