--- x/feedist/msg_test.go
@@ -1,1 +1,0 @@
-package feedist
